+++ backend/src/lambda/http/createTodo.ts
@@ -1,41 +1,15 @@
 import 'source-map-support/register'
-import * as AWS  from 'aws-sdk'
 import { APIGatewayProxyEvent, APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda'
-import { CreateTodoRequest } from '../../requests/CreateTodoRequest'
-import * as uuid from 'uuid'
-import { parseUserId } from '../../auth/utils'
 import {createLogger} from "../../utils/logger";
-
-const docClient = new AWS.DynamoDB.DocumentClient()
-const todosTable = process.env.TODOS_TABLE
+import {createTodo} from "../../services/manageTodos"
 
 const logger = createLogger('createToDos')
 
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
 
   logger.info('Processing event: ', event)
-  const itemId = uuid.v4()
 
-  const newTodo: CreateTodoRequest = JSON.parse(event.body)
-
-  const authorization = event.headers.Authorization
-  const split = authorization.split(' ')
-  const jwtToken = split[1]
-
-  const item = {
-    todoId: itemId,
-    userId: parseUserId(jwtToken),
-    createdAt: new Date(Date.now()).toISOString(),
-    done: false,
-    ...newTodo
-  }
-
-
-
-  await docClient.put({
-    TableName: todosTable,
-    Item: item
-  }).promise()
+  const item = await createTodo(event)
 
   return {
     statusCode: 201,
