+++ backend/src/lambda/http/deleteTodo.ts
@@ -1,48 +1,39 @@
 import 'source-map-support/register'
 import { APIGatewayProxyEvent, APIGatewayProxyResult, APIGatewayProxyHandler } from 'aws-lambda'
-import * as AWS from "aws-sdk";
-import {parseUserId} from "../../auth/utils";
 import {createLogger} from "../../utils/logger";
-
-const docClient = new AWS.DynamoDB.DocumentClient()
-const todosTable = process.env.TODOS_TABLE
+import {deleteTodo} from "../../services/manageTodos"
 
 const logger = createLogger('createToDos')
 
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
 
   logger.info('Processing event: ', event)
-  const todoId = event.pathParameters.todoId
 
-  const authorization = event.headers.Authorization
-  const split = authorization.split(' ')
-  const jwtToken = split[1]
+  try {
+    await deleteTodo(event)
 
-  const result = await docClient.get({
-    TableName: todosTable,
-    Key: {
-      todoId,
-      userId: parseUserId(jwtToken)
-    }
-  }).promise();
+    return {
+      statusCode: 202,
+      headers: {
+        'Access-Control-Allow-Origin': '*',
+        'Access-Control-Allow-Credentials': true
+      },
+      body: JSON.stringify({})
+    };
 
+  } catch (e) {
 
-  if (result.Item.size <=0) {
     return {
       statusCode: 404,
       body: JSON.stringify({
-        error: 'Item does not exist'
+        error: e
       })
     };
+
   }
 
-  await docClient.delete({
-    TableName: todosTable,
-    Key: {
-      todoId,
-      userId: parseUserId(jwtToken)
-    }
-  }).promise();
+
+
 
   return {
     statusCode: 202,
