+++ backend/src/lambda/http/generateUploadUrl.ts
@@ -1,50 +1,17 @@
 import 'source-map-support/register'
 import { APIGatewayProxyEvent, APIGatewayProxyResult, APIGatewayProxyHandler } from 'aws-lambda'
-import * as AWS  from 'aws-sdk'
-import {parseUserId} from "../../auth/utils";
-import {createLogger} from "../../utils/logger";
 
-const s3 = new AWS.S3({
-  signatureVersion: 'v4'
-})
+import {createLogger} from "../../utils/logger";
+import {uploadImage} from "../../services/manageTodos"
+
 
 const logger = createLogger('generateUploadURL')
-
-const bucketName = process.env.ATTACHMENTS_S3_BUCKET
-const urlExpiration = process.env.SIGNED_URL_EXPIRATION
-const docClient = new AWS.DynamoDB.DocumentClient()
-const todosTable = process.env.TODOS_TABLE
 
 export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
 
   logger.info('Processing event: ', event)
-  const todoId = event.pathParameters.todoId
 
-  const authorization = event.headers.Authorization
-  const split = authorization.split(' ')
-  const jwtToken = split[1]
-
-  const createSignedUrlRequest = {
-    Bucket: bucketName,
-    Key: todoId,
-    Expires: urlExpiration
-  }
-
-  const signedUrl = s3.getSignedUrl('putObject', createSignedUrlRequest)
-  const attachmentUrl = `https://${bucketName}.s3.amazonaws.com/${todoId}`
-
-  await docClient.update({
-    TableName: todosTable,
-    Key: {
-      todoId,
-      userId: parseUserId(jwtToken)
-    },
-    UpdateExpression: 'set attachmentUrl = :a',
-    ExpressionAttributeValues: {
-      ':a': attachmentUrl
-    },
-    ReturnValues:"UPDATED_NEW"
-  }).promise();
+  const signedUrl = await uploadImage(event)
 
   return {
     statusCode: 202,
