+++ backend/src/lambda/http/getTodos.ts
@@ -1,13 +1,8 @@
 import { APIGatewayProxyHandler, APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda'
 import 'source-map-support/register'
-import * as AWS  from 'aws-sdk'
-import { parseUserId } from '../../auth/utils'
-import {createLogger} from "../../utils/logger";
+import {createLogger} from "../../utils/logger"
+import {getTodo} from "../../services/manageTodos"
 
-const docClient = new AWS.DynamoDB.DocumentClient()
-
-const todosTable = process.env.TODOS_TABLE
-const todosIdIndex = process.env.TODOS_ID_INDEX
 
 const logger = createLogger('getToDos')
 
@@ -15,20 +10,7 @@ export const handler: APIGatewayProxyHandler = async (event: APIGatewayProxyEven
 
     logger.info('Processing event: ', event)
 
-    const authorization = event.headers.Authorization
-    const split = authorization.split(' ')
-    const jwtToken = split[1]
-
-    const result = await docClient.query({
-        TableName: todosTable,
-        IndexName: todosIdIndex,
-        KeyConditionExpression: 'userId = :userId',
-        ExpressionAttributeValues: {
-            ':userId': parseUserId(jwtToken)
-        }
-    }).promise();
-
-    const items = result.Items
+    const items = await getTodo(event)
 
     return {
         statusCode: 200,
